BOL-375 refactor: login package renaming to auth (api와 맞춤)
