+++ data/src/main/java/com/yapp/bol/data/datasource/user/UserDataSource.kt
@@ -1,8 +1,8 @@
 package com.yapp.bol.data.datasource.user
 
 import com.yapp.bol.data.model.group.response.JoinedGroupResponse
-import com.yapp.bol.data.model.login.OnBoardResponse
-import com.yapp.bol.data.model.login.UserRequest
+import com.yapp.bol.data.model.auth.OnBoardResponse
+import com.yapp.bol.data.model.auth.UserRequest
 import com.yapp.bol.data.model.user.UserResponse
 import com.yapp.bol.domain.model.ApiResult
 import kotlinx.coroutines.flow.Flow
