+++ data/src/main/java/com/yapp/bol/data/model/auth/TermsDTO.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.data.model.login
+package com.yapp.bol.data.model.auth
 
 data class TermsDTO(
     val code: String,
