+++ data/src/main/java/com/yapp/bol/data/model/auth/UserRequest.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.data.model.login
+package com.yapp.bol.data.model.auth
 
 data class UserRequest(
     val nickname: String
