+++ data/src/main/java/com/yapp/bol/data/remote/AuthApi.kt
@@ -1,7 +1,7 @@
 package com.yapp.bol.data.remote
 
-import com.yapp.bol.data.model.login.LoginRequest
-import com.yapp.bol.data.model.login.LoginResponse
+import com.yapp.bol.data.model.auth.LoginRequest
+import com.yapp.bol.data.model.auth.LoginResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
