+++ data/src/main/java/com/yapp/bol/data/remote/TermsApi.kt
@@ -1,7 +1,7 @@
 package com.yapp.bol.data.remote
 
-import com.yapp.bol.data.model.login.TermsRequest
-import com.yapp.bol.data.model.login.TermsResponse
+import com.yapp.bol.data.model.auth.TermsRequest
+import com.yapp.bol.data.model.auth.TermsResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
